Moved SONiC environment logic into a separate SoniCEnvironment class

+++ common/sai_client/sai_redis_client/sai_redis_client.py
@@ -7,44 +7,105 @@ import paramiko
 from saichallenger.common.sai_client.sai_client import SaiClient
 from saichallenger.common.sai_data import SaiObjType, SaiData
 
-def service_is_active(target, service):
-    _, stdout, _ = target.exec_command(f"systemctl is-active {service}")
-    output = stdout.read().decode("utf-8")
-    return "inactive" not in output
 
+class SonicEnvironment():
+    def __init__(self, cfg):
+        self.server_ip = cfg["ip"]
+        self.port = cfg["port"]
 
-def container_is_running(target, container):
-    _, stdout, _ = target.exec_command(f"docker inspect {container}")
-    output = stdout.read().decode("utf-8")
-    return json.loads(output)[0]["State"]["Running"]
+        self.ssh = paramiko.SSHClient()
+        self.ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+        self.ssh.connect(
+            self.server_ip,
+            username=cfg.get("username", "admin"),
+            password=cfg.get("password", "YourPaSsWoRd")
+        )
 
-def assert_container_state(target, container, is_running=True, tout=30):
-    for i in range(tout):
-        time.sleep(1)
-        if container_is_running(target, container) == is_running:
-            return
-    state = "not running" if is_running else "running"
-    assert False, f"The {container} container is still not running after {tout} seconds..."
+        # In SONiC environment, Redis is listening on loopback interface only.
+        # So, we can retrieve device metadata though SSH only.
+        _, stdout, _ = self.ssh.exec_command('redis-cli -n 4 --raw hgetall "DEVICE_METADATA|localhost"')
+        output = stdout.readlines()
+        device_metadata = {}
+        for i in range(0, len(output), 2):
+            device_metadata[output[i][:-2]] = output[i + 1][:-2]
 
-def assert_service_state(target, service, is_active=True, tout=30):
-    for i in range(tout):
-        time.sleep(1)
-        if service_is_active(target, service) == is_active:
-            return
-    state = "inactive" if is_active else "active"
-    assert False, f"The {service} service is still {state} after {tout} seconds..."
+        # Enable Redis server to listen on all interfaces
+        cmd = "echo \"sed -ri 's/--bind.*--port/--bind 0.0.0.0 --port/' /usr/share/sonic/templates/supervisord.conf.j2\" > redis_bind_fix.sh"
+        self.ssh.exec_command(cmd)
+        self.ssh.exec_command("docker cp redis_bind_fix.sh database:/")
+        self.ssh.exec_command("docker exec database bash redis_bind_fix.sh")
 
-def assert_redis_is_available(host, port, tout=30):
-    start_time = time.time()
-    r = redis.Redis(host=host, port=port, db=0)
-    while True:
-        try:
-            r.ping()
-            return
-        except:
-            if time.time() - start_time < tout:
-                time.sleep(1)
-                continue
+        # Stop all SONiC services
+        for service in ["monit", "pmon", "sonic.target", "syncd", "swss", "database"]:
+            self.ssh.exec_command(f"sudo systemctl mask {service}")
+            self.ssh.exec_command(f"sudo systemctl stop {service}")
+            self._assert_service_state(service, is_active=False, tout=60)
+
+        # Stop SyncD just in case it's the second run of SAI-C
+        self.ssh.exec_command("docker stop syncd")
+        self._assert_container_state("syncd", is_running=False)
+
+        # Apply Redis config change
+        self.ssh.exec_command("docker stop database")
+        self._assert_container_state("database", is_running=False)
+        self.ssh.exec_command("docker start database")
+        self._assert_container_state("database", is_running=True)
+        self._assert_redis_is_available()
+
+        # Flush SONiC Redis content
+        r = redis.Redis(host=self.server_ip, port=self.port, db=1)
+        r.flushall()
+
+        # Write to CONFIG_DB SONiC device information needed on syncd start
+        config_db = redis.Redis(host=self.server_ip, port=self.port, db=4)
+        config_db.hmset("DEVICE_METADATA|localhost", device_metadata)
+        config_db.set("CONFIG_DB_INITIALIZED", "1")
+
+    def cleanup(self):
+        self.ssh.exec_command("docker stop syncd")
+        self._assert_container_state("syncd", is_running=False)
+        # Flush ASIC_DB content
+        r = redis.Redis(host=self.server_ip, port=self.port, db=1)
+        r.flushdb()
+        self.ssh.exec_command("docker start syncd")
+
+    def _service_is_active(self, service):
+        _, stdout, _ = self.ssh.exec_command(f"systemctl is-active {service}")
+        output = stdout.read().decode("utf-8")
+        return "inactive" not in output
+
+    def _container_is_running(self, container):
+        _, stdout, _ = self.ssh.exec_command(f"docker inspect {container}")
+        output = stdout.read().decode("utf-8")
+        return json.loads(output)[0]["State"]["Running"]
+
+    def _assert_container_state(self, container, is_running=True, tout=30):
+        for i in range(tout):
+            time.sleep(1)
+            if self._container_is_running(container) == is_running:
+                return
+        state = "not running" if is_running else "running"
+        assert False, f"The {container} container is still not running after {tout} seconds..."
+
+    def _assert_service_state(self, service, is_active=True, tout=30):
+        for i in range(tout):
+            time.sleep(1)
+            if self._service_is_active(service) == is_active:
+                return
+        state = "inactive" if is_active else "active"
+        assert False, f"The {service} service is still {state} after {tout} seconds..."
+
+    def _assert_redis_is_available(self, tout=30):
+        start_time = time.time()
+        r = redis.Redis(host=self.server_ip, port=self.port, db=0)
+        while True:
+            try:
+                r.ping()
+                return
+            except:
+                if time.time() - start_time < tout:
+                    time.sleep(1)
+                    continue
             assert False, f"Redis server is still not available after {tout} seconds..."
 
 
@@ -59,53 +120,16 @@ class SaiRedisClient(SaiClient):
         self.port = cfg["port"]
         self.libsaivs = cfg["saivs"]
 
-        self.ssh = paramiko.SSHClient()
-        self.ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        self.ssh.connect(self.server_ip, username="admin", password="YourPaSsWoRd")
-
-        _, stdout, _ = self.ssh.exec_command('redis-cli -n 4 --raw hgetall "DEVICE_METADATA|localhost"')
-        output = stdout.readlines()
-        self.device_metadata = {}
-        for i in range(0, len(output), 2):
-            self.device_metadata[output[i][:-2]] = output[i + 1][:-2]
-
-        # Enable Redis server to listen on all interfaces
-        cmd = "echo \"sed -ri 's/--bind.*--port/--bind 0.0.0.0 --port/' /usr/share/sonic/templates/supervisord.conf.j2\" > redis_bind_fix.sh"
-        self.ssh.exec_command(cmd)
-        self.ssh.exec_command("docker cp redis_bind_fix.sh database:/")
-        self.ssh.exec_command("docker exec database bash redis_bind_fix.sh")
-
-        # Stop all SONiC services
-        for service in ["monit", "pmon", "sonic.target", "syncd", "swss", "database"]:
-            self.ssh.exec_command(f"sudo systemctl mask {service}")
-            self.ssh.exec_command(f"sudo systemctl stop {service}")
-            assert_service_state(self.ssh, service, is_active=False, tout=60)
-
-        # Stop SyncD just in case it's the second run of SAI-C
-        self.ssh.exec_command("docker stop syncd")
-        assert_container_state(self.ssh, "syncd", is_running=False)
-
-        # Apply Redis config change
-        self.ssh.exec_command("docker stop database")
-        assert_container_state(self.ssh, "database", is_running=False)
-        self.ssh.exec_command("docker start database")
-        assert_container_state(self.ssh, "database", is_running=True)
-        assert_redis_is_available(self.server_ip, self.port)
+        self.env = None
+        if cfg.get("mode", None) == "sonic":
+            self.env = SonicEnvironment(cfg)
 
         self.r = redis.Redis(host=self.server_ip, port=self.port, db=1)
         self.loglevel_db = redis.Redis(host=self.server_ip, port=self.port, db=3)
-        self.config_db = redis.Redis(host=self.server_ip, port=self.port, db=4)
-        self.r.flushall()
-
-        # Write to CONFIG_DB SONiC device information needed on syncd start
-        self.config_db.hmset("DEVICE_METADATA|localhost", self.device_metadata)
-        self.config_db.set("CONFIG_DB_INITIALIZED", "1")
-
         self.cache = {}
         self.rec2vid = {}
 
         self.switch_oid = "oid:0x21000000000000"
-
 
     def cleanup(self):
         '''
@@ -115,8 +139,16 @@ class SaiRedisClient(SaiClient):
         the system (DUT) is in the initial state with no extra SAI objects
         created. To ensure this, the framework should flush Redis DB content
         and restart syncd application linked with SAI library.
+        '''
+        if self.env:
+            self.env.cleanup()
+            self.cache = {}
+            self.rec2vid = {}
+            self.__assert_syncd_running()
+            return
 
-        The execution flow:
+        '''
+        SAI-C server environment's flow:
           1. On Docker start, the supervisord starts Redis server and syncd.
              For more details, please see `supervisord.conf` file.
           2. This function flushes Redis DB content by FLUSHALL command
@@ -128,19 +160,12 @@ class SaiRedisClient(SaiClient):
           5. The supervisord restarts syncd program as per `autorestart`
              option in `supervisord.conf` file.
         '''
-        #self.r.flushall()
-        #self.loglevel_db.hmset('syncd:syncd', {'LOGLEVEL':self.loglevel, 'LOGOUTPUT':'SYSLOG'})
-        #self.r.shutdown()
-        #time.sleep(2)
-        #self.cache = {}
-        #self.rec2vid = {}
-        self.ssh.exec_command("docker stop syncd")
-        assert_container_state(self.ssh, "syncd", is_running=False)
-        self.r.flushdb()
-        self.ssh.exec_command("docker start syncd")
+        self.r.flushall()
+        self.loglevel_db.hmset('syncd:syncd', {'LOGLEVEL':self.loglevel, 'LOGOUTPUT':'SYSLOG'})
+        self.r.shutdown()
+        time.sleep(2)
         self.cache = {}
         self.rec2vid = {}
-
         self.__assert_syncd_running()
 
     def set_loglevel(self, sai_api, loglevel):
@@ -532,12 +557,15 @@ class SaiRedisClient(SaiClient):
 
     # Host interface
     def remote_iface_exists(self, iface):
+        assert self.env, "Operation is not supported in SONiC environment"
         return self.__remote_cmd_operate("iface_exists", iface) == "ok"
 
     def remote_iface_is_up(self, iface):
+        assert self.env, "Operation is not supported in SONiC environment"
         return self.__remote_cmd_operate("iface_is_up", iface) == "ok"
 
     def remote_iface_status_set(self, iface, status):
+        assert self.env, "Operation is not supported in SONiC environment"
         admin = "up" if status else "down"
         args = {
             "iface": iface,
@@ -546,9 +574,11 @@ class SaiRedisClient(SaiClient):
         return self.__remote_cmd_operate("set_iface_status", args) == "ok"
 
     def remote_iface_agent_start(self, ifaces):
+        assert self.env, "Operation is not supported in SONiC environment"
         return self.__remote_cmd_operate("start_nn_agent", ifaces) == "ok"
 
     def remote_iface_agent_stop(self):
+        assert self.env, "Operation is not supported in SONiC environment"
         return self.__remote_cmd_operate("stop_nn_agent") == "ok"
 
     def get_oids(self, obj_type=None):
